requirements.txt
asgiref==3.7.2
certifi==2023.11.17
charset-normalizer==3.3.2
Django==4.2.7
django-cors-headers==4.3.1
django-dirtyfields==1.9.2
djangorestframework==3.14.0
idna==3.4
pyTelegramBotAPI==4.14.0
pytz==2023.3.post1
requests==2.31.0
sqlparse==0.4.4
telebot==0.0.5
typing_extensions==4.8.0
urllib3==2.1.0
